edit customer analytics

+++ scripts/04_customer_analytics.sql
@@ -1,7 +1,7 @@
 -- Revenue by gender
 
 SELECT u.gender,
-ROUND(SUM(p.total_amount::integer),2) as total_revenue
+ROUND(SUM(p.total_amount::numeric),2) as total_revenue
 FROM purchases as p
 JOIN users as u
 ON p.user_id = u.user_id
@@ -11,7 +11,7 @@ ORDER BY total_revenue DESC;
 -- Revenue by country
 
 SELECT u.country,
-ROUND(SUM(p.total_amount::integer), 2) as total_revenue
+ROUND(SUM(p.total_amount::numeric), 2) as total_revenue
 FROM purchases as p
 JOIN users as u
 ON p.user_id = u.user_id
@@ -21,7 +21,7 @@ ORDER BY total_revenue DESC;
 -- Top customers
 
 SELECT u.user_id,
-ROUND(SUM(p.total_amount::integer), 2) as revenue
+ROUND(SUM(p.total_amount::numeric), 2) as revenue
 FROM purchases as p
 JOIN users as u
 ON p.user_id = u.user_id
